Test: TestStore를 파일로 분리
TestStore는 reducer를 테스트하기위한 테스트용 모듈이다. 그렇기때문에
따로 분리해서 다른 테스트 코드도 사용할 수 있게 만들었다. 추가적으로
async한 reducer 테스트를 위해서 async dispatch도 지원하게 만들었다.
